refactor: move offre & immersions saved search actions
