+++ lib/features/immersion/list/immersion_list_actions.dart
@@ -1,4 +1,5 @@
 import 'package:pass_emploi_app/models/immersion.dart';
+import 'package:pass_emploi_app/models/immersion_filtres_parameters.dart';
 import 'package:pass_emploi_app/models/location.dart';
 
 class ImmersionListRequestAction {
@@ -20,3 +21,15 @@ class ImmersionListSuccessAction {
 class ImmersionListFailureAction {}
 
 class ImmersionListResetAction {}
+
+class ImmersionSearchParametersRequestAction {
+  final String codeRome;
+  final Location location;
+  final ImmersionSearchParametersFiltres filtres;
+
+  ImmersionSearchParametersRequestAction({
+    required this.codeRome,
+    required this.location,
+    required this.filtres,
+  });
+}
