+++ lib/features/offre_emploi/parameters/offre_emploi_search_parameters_actions.dart
@@ -1,20 +1,5 @@
-import 'package:pass_emploi_app/models/location.dart';
 import 'package:pass_emploi_app/models/offre_emploi.dart';
 import 'package:pass_emploi_app/models/offre_emploi_filtres_parameters.dart';
-
-class OffreEmploiSearchParametersRequestAction {
-  final String keywords;
-  final Location? location;
-  final bool onlyAlternance;
-  final OffreEmploiSearchParametersFiltres filtres;
-
-  OffreEmploiSearchParametersRequestAction({
-    required this.keywords,
-    required this.location,
-    required this.onlyAlternance,
-    required this.filtres,
-  });
-}
 
 class OffreEmploiSearchParametersUpdateFiltresRequestAction {
   final OffreEmploiSearchParametersFiltres updatedFiltres;
